Split spreadinterp into spread, interp, and spreadinterp detail headers
Extract the heavy SIMD-templated spread/interp code from the monolithic
spreadinterp.cpp into separate detail headers and per-precision TUs:
- detail/simd_helpers.hpp: shared SIMD helpers, fold_rescale, evaluate_kernel_vector
- detail/spread.hpp: spread_subproblem_*, spreadSorted
- detail/interp.hpp: interp_line/square/cube, interpSorted
- detail/spreadinterp.hpp: indexSort, spreadinterpSorted, evaluate_kernel_runtime
- src/spread.cpp / src/interp.cpp: explicit instantiation TUs
- src/spreadinterp.cpp: extern template + remaining instantiations
- spreadcheck() moved inline into public spreadinterp.hpp header
